Refactor imports to use absolute paths and change order of codecheck workflow
