+++ .github/workflows/codecheck.yaml
@@ -31,8 +31,8 @@ jobs:
           pip install ruff
           pip list
 
-      - name: Ruff check
-        run: ruff check .
-
       - name: Ruff format
         run: ruff format .
+
+      - name: Ruff check
+        run: ruff check .
