+++ src/face_classification/evaluate.py
@@ -11,8 +11,8 @@ from pytorch_lightning import Trainer
 from torch.utils.data import DataLoader
 from pytorch_lightning.loggers import WandbLogger
 
-from face_classification.data import FaceDataset
-from face_classification.model import PretrainedResNet34
+from src.face_classification.data import FaceDataset
+from src.face_classification.model import PretrainedResNet34
 
 app = typer.Typer()
 @app.command()
