+++ src/face_classification/model.py
@@ -6,7 +6,7 @@ from torchmetrics import Accuracy
 from torchvision import models
 from torchvision.models import ResNet34_Weights
 from omegaconf import DictConfig
-from face_classification.utils import accuracy
+from src.face_classification.utils import accuracy
 
 # Optimizer options
 OPTIMIZER_OPTIONS = {"adam": torch.optim.Adam, "sgd": torch.optim.SGD}
