+++ src/face_classification/train.py
@@ -9,9 +9,9 @@ from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import WandbLogger
 from torch.profiler import profile, ProfilerActivity, tensorboard_trace_handler
 
-from face_classification.metric_tracker import MetricTracker
-from face_classification.model import PretrainedResNet34
-from face_classification.data import FaceDataset
+from src.face_classification.metric_tracker import MetricTracker
+from src.face_classification.model import PretrainedResNet34
+from src.face_classification.data import FaceDataset
 import os
 
 app = typer.Typer()
